requirements.txt
pandas>=2.2.0
numpy>=1.26.0
xgboost>=2.0.0
scikit-learn>=1.4.0
statsmodels>=0.14.0
scipy>=1.13.0
pyarrow>=15.0.0
openpyxl>=3.1.0
xlrd>=2.0.1
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.6.0
pydantic[email]>=2.6.0
httpx>=0.27.0
python-dotenv>=1.0.0
streamlit>=1.33.0
plotly>=5.20.0
requests>=2.31.0
jinja2>=3.1.2
APScheduler>=3.10.4